requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
importlib-metadata==4.11.4
Markdown==3.3.7
numpy==1.23.0
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0
